+++ src/main/java/chess/ChessApplication.java
@@ -2,7 +2,7 @@ package chess;
 
 import chess.controller.ChessController;
 
-public class ChessApplication {
+public final class ChessApplication {
     public static void main(String[] args) {
         new ChessController().run();
     }
